+++ src/core/hooks/interfaces/HookHandler.ts
@@ -1,4 +1,4 @@
-import { AllHooks, HookKeys } from '../types/HookMap';
+import { AllHooks, HookKeys } from '../types/Hooks';
 
 export abstract class HookHandler<T extends HookKeys> {
   protected data: AllHooks[T];
