+++ src/core/hooks/interfaces/abstracts/WebhookLog.ts
@@ -1,6 +1,6 @@
 import { WebhookClient } from 'discord.js';
 import { HookHandler } from '../HookHandler';
-import { HookKeys } from '../../types/HookMap';
+import { HookKeys } from '../../types/Hooks';
 
 export abstract class WebhookLog<T extends HookKeys> extends HookHandler<T> {
   abstract webhook: WebhookClient;
